Chiron AI, Business Plan - MBA Milestone Project
This is a business concept developed as part of the MBA program at the University of Illinois Urbana-Champaign, Gies College of Business. It represents a milestone project and is not yet in production. The concept is designed to be expanded and refined as technology, funding, and partnerships make full implementation feasible.